Implement JVMTI GetPhase.

+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/ffi/nfi/NativeUtils.java
@@ -110,7 +110,7 @@ public final class NativeUtils {
         if (library.isNull(pointer)) {
             return;
         }
-        IntBuffer resultPointer = NativeUtils.directByteBuffer(pointer, 1, JavaKind.Byte).asIntBuffer();
+        IntBuffer resultPointer = NativeUtils.directByteBuffer(pointer, 1, JavaKind.Int).asIntBuffer();
         resultPointer.put(value);
     }
